refactor: move client list into pages
