+++ src/components/index.ts
@@ -7,5 +7,5 @@ export { ToggleSwitch } from './ToggleSwitch';
 export { ToolsModal } from './ToolsModal';
 export { LoadingButton } from './LoadingButton';
 export { Toast, ToastContainer, type ToastType } from './Toast';
-export { ClientList } from './ClientList';
+export { ClientCard } from './clients/ClientCard';
 export { TabSwitcher, type TabKey } from './TabSwitcher';
